[Refactor] split client into single file
